+++ examples/web/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @podley/cli
 
+## 0.0.29
+
+### Patch Changes
+
+- Cleanup GraphAsTask input/output schemas
+- Updated dependencies
+  - @podley/ai-provider@0.0.29
+  - @podley/task-graph@0.0.29
+  - @podley/job-queue@0.0.29
+  - @podley/storage@0.0.29
+  - @podley/sqlite@0.0.29
+  - @podley/tasks@0.0.29
+  - @podley/test@0.0.29
+  - @podley/util@0.0.29
+  - @podley/ai@0.0.29
+
+## 0.0.28
+
+### Patch Changes
+
+- inputSchema and outputSchema dynamically created for GraphAsTask
+- 6915093: inputSchema and outputSchema for GraphAsTask
+- Updated dependencies [6915093]
+- Updated dependencies
+- Updated dependencies [6915093]
+  - @podley/ai-provider@0.0.28
+  - @podley/ai@0.0.28
+  - @podley/task-graph@0.0.28
+  - @podley/test@0.0.28
+  - @podley/job-queue@0.0.28
+  - @podley/sqlite@0.0.28
+  - @podley/storage@0.0.28
+  - @podley/tasks@0.0.28
+  - @podley/util@0.0.28
+
 ## 0.0.20
 
 ### Patch Changes