+++ packages/util/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @podley/util
 
+## 0.0.29
+
+### Patch Changes
+
+- Cleanup GraphAsTask input/output schemas
+
+## 0.0.28
+
+### Patch Changes
+
+- inputSchema and outputSchema dynamically created for GraphAsTask
+- 6915093: inputSchema and outputSchema for GraphAsTask
+
 ## 0.0.20
 
 ## 0.0.19